clean up useEffects for content search
